complete(web): Done

+++ spotify-server/src/services/song.service.js
@@ -1,5 +1,7 @@
 const Song = require('../models/song.model')
+const Album = require('../models/album.model')
 const { uploadFile } = require('../util/uploadFile.util')
+const { default: mongoose } = require('mongoose')
 
 const getFormattedDuration = (duration) => {
     const minutes = Math.floor(duration / 60).toString().padStart(2, '0')
@@ -9,17 +11,20 @@ const getFormattedDuration = (duration) => {
 
 const handleAddSong = async ({name, description, album, audioFile, imageFile}) => {
     try {
+        const existingAlbum = await Album.findById(album)
+        if (!existingAlbum) throw new Error("Album not found")
         const [audioUpload, imageUpload] = await Promise.all([
-            uploadFile(audioFile, 'video'),
-            uploadFile(imageFile, 'image')
-        ])
+            uploadFile(audioFile, "auto"),
+            uploadFile(imageFile, "auto"),
+        ]);
+
 
         const duration = getFormattedDuration(audioUpload.duration)
 
         const songData = {
             name, 
             description,
-            album,
+            album: new mongoose.Types.ObjectId(album),
             image: imageUpload.secure_url,
             file: audioUpload.secure_url,
             duration
@@ -27,21 +32,48 @@ const handleAddSong = async ({name, description, album, audioFile, imageFile}) =
 
         const newSong = new Song(songData)
         await newSong.save()
+        existingAlbum.songs.push(newSong._id)
+        await existingAlbum.save()
+
+        await newSong.populate('album', '_id name');
 
         return newSong
     } catch (error) {
-        console.error(">> Error when adding song: ", error.message)
-        return null
+        console.error(">> Error when adding song: ", error)
+        throw error
     }
 }
 
 const handleGetListSong = async () => {
-    const data = await Song.find({})
-    return data
+    try {
+        const data = await Song.find({}).populate('album', '_id name')
+        return data
+    } catch (error) {
+        console.error("Error loading song", error)
+        throw error
+    }
 }
 
 const handleDeleteSong = async ( id ) => {
-    return await Song.findByIdAndDelete(id)
+    try {
+        const song = await Song.findById(id)
+        if (!song) throw new Error("Song Id not found")
+        
+        await Song.findByIdAndDelete(id)
+        if (Song.album) {
+            await Album.findByIdAndUpdate(song.album, {
+                $pull: { songs: song._id }
+            })
+        }
+
+        return {
+            success: true,
+            message: "Song removed successfully"
+        }
+    } catch (error) {
+        console.error("Error deleting song: ", error)
+        throw error
+    }
 }
 
 module.exports = {
